chore(ci) Utilize new upstream Luacheck Action
As of v0.26.0, the Luacheck repository is enabled to run *as* a GH
Action so there is no need to install Lua, LuaRocks, then the linter,
etc. Just one `uses:`.

This also removes Luarocks rockspec linter, but I think that belongs in
an altogether different place in the workflow(s).